backend changes

+++ database/migrations/2023_03_09_200125_create_hearing_tests_table.php
@@ -17,7 +17,7 @@ class CreateHearingTestsTable extends Migration
             $table->id();
             $table->integer('sound_id');
             $table->integer('sound_volume');
-            $table->integer('score');
+            $table->decimal('score',3,2);
             $table->integer('test_id');
             $table->timestamps();
         });
