+++ database/migrations/2023_03_09_211122_create_hearing_test_results_table.php
@@ -16,7 +16,7 @@ class CreateHearingTestResultsTable extends Migration
         Schema::create('hearing_test_results', function (Blueprint $table) {
             $table->id();
             $table->integer('test_id');
-            $table->integer('score');
+            $table->decimal('score',3,2);
             $table->timestamps();
         });
     }
